Fix register allocation
Right now, it's possible to clobber a temporary register before we should
when a predicate with a lower arity calls one with a higher arity and no
permanent registers are involved. This fix takes care of that by starting
the register allocation on the lowest temporary register that is unused by
the predicate's arguments and all of its callees.